# src/main/resources/webdoc/openapi.yaml
openapi: 3.0.3
info:
  title: CommunityBudget API
  version: 1.0.0
servers:
  - url: http://localhost:8080

tags:
  - name: User
    description: Operaciones relacionadas con usuarios
  - name: Auth
    description: Operaciones de autenticación

security:
  - BearerAuth: [ ]

paths:
  /auth/login:
    post:
      operationId: login
      tags:
        - Auth
      summary: Iniciar sesión
      description: Autentica al usuario y devuelve un token JWT
      security: [ ]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequestDTO'
      responses:
        '200':
          description: Login exitoso
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LoginResponseDTO'
        '401':
          description: Credenciales inválidas

  /auth/refresh:
    post:
      operationId: refreshToken
      tags:
        - Auth
      summary: Refrescar token
      description: Obtiene un nuevo token JWT usando el refresh token
      security: [ ]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RefreshTokenRequestDTO'
      responses:
        '200':
          description: Token refrescado exitosamente
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LoginResponseDTO'
        '401':
          description: Refresh token inválido o expirado

  /auth/register:
    post:
      operationId: register
      tags:
        - Auth
      summary: Registrar un nuevo usuario
      description: Registra un nuevo usuario en el sistema
      security: [ ]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserCreateDTO'
      responses:
        '201':
          description: Usuario creado exitosamente
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserDTO'
        '400':
          description: Datos de entrada inválidos
        '409':
          description: El usuario ya existe

  /user/me:
    get:
      operationId: getCurrentUser
      tags:
        - User
      summary: Obtener información del usuario autenticado
      description: Retorna la información del usuario actualmente autenticado
      responses:
        '200':
          description: Usuario obtenido exitosamente
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserDTO'
        '401':
          description: No autorizado
        '404':
          description: Usuario no encontrado

    put:
      operationId: updateCurrentUser
      tags:
        - User
      summary: Actualizar usuario autenticado
      description: Actualiza la información del usuario autenticado
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserUpdateDTO'
      responses:
        '200':
          description: Usuario actualizado exitosamente
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserDTO'
        '400':
          description: Datos de entrada inválidos
        '401':
          description: No autorizado

    delete:
      operationId: deleteCurrentUser
      tags:
        - User
      summary: Eliminar usuario autenticado
      description: Elimina la cuenta del usuario autenticado
      responses:
        '204':
          description: Usuario eliminado exitosamente
        '401':
          description: No autorizado

  /user/{id}:
    get:
      operationId: getUserById
      tags:
        - User
      summary: Obtener usuario por ID
      description: Retorna la información de un usuario específico
      parameters:
        - name: id
          in: path
          required: true
          description: ID del usuario
          schema:
            type: string
      responses:
        '200':
          description: Usuario obtenido exitosamente
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserDTO'
        '404':
          description: Usuario no encontrado

    put:
      operationId: updateUserById
      tags:
        - User
      summary: Actualizar usuario por ID
      description: Actualiza la información de un usuario existente (solo admin)
      parameters:
        - name: id
          in: path
          required: true
          description: ID del usuario
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserUpdateDTO'
      responses:
        '200':
          description: Usuario actualizado exitosamente
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserDTO'
        '400':
          description: Datos de entrada inválidos
        '404':
          description: Usuario no encontrado

    delete:
      operationId: deleteUserById
      tags:
        - User
      summary: Eliminar usuario por ID
      description: Elimina un usuario del sistema (solo admin)
      parameters:
        - name: id
          in: path
          required: true
          description: ID del usuario
          schema:
            type: string
      responses:
        '204':
          description: Usuario eliminado exitosamente
        '404':
          description: Usuario no encontrado

  /user/change-password:
    put:
      operationId: changePassword
      tags:
        - User
      summary: Cambiar contraseña
      description: Permite al usuario autenticado cambiar su contraseña
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ChangePasswordDTO'
      responses:
        '200':
          description: Contraseña actualizada exitosamente
        '400':
          description: Datos de entrada inválidos o contraseña actual incorrecta
        '401':
          description: No autorizado

components:
  schemas:
    UserDTO:
      type: object
      properties:
        id:
          type: string
          description: Identificador único del usuario
        name:
          type: string
          description: Nombre del usuario
        email:
          type: string
          format: email
          description: Correo electrónico del usuario
        avatarUrl:
          type: string
          description: URL del avatar del usuario

    UserCreateDTO:
      type: object
      required:
        - name
        - email
        - password
      properties:
        name:
          type: string
          description: Nombre del usuario
          minLength: 2
          maxLength: 100
        email:
          type: string
          format: email
          description: Correo electrónico del usuario
        password:
          type: string
          format: password
          description: Contraseña del usuario (mínimo 8 caracteres, debe incluir mayúscula, minúscula, número y carácter especial)
          minLength: 8
          maxLength: 128
          pattern: '^(?=.*[a-z])(?=.*[A-Z])(?=.*\d)(?=.*[@$!%*?&])[A-Za-z\d@$!%*?&]{8,}$'
        avatarUrl:
          type: string
          description: URL del avatar del usuario

    UserUpdateDTO:
      type: object
      properties:
        name:
          type: string
          description: Nombre del usuario
          minLength: 2
          maxLength: 100
        email:
          type: string
          format: email
          description: Correo electrónico del usuario
        avatarUrl:
          type: string
          description: URL del avatar del usuario

    LoginRequestDTO:
      type: object
      required:
        - email
        - password
      properties:
        email:
          type: string
          format: email
          description: Correo electrónico del usuario
        password:
          type: string
          format: password
          description: Contraseña del usuario

    LoginResponseDTO:
      type: object
      properties:
        accessToken:
          type: string
          description: Token JWT de acceso
        refreshToken:
          type: string
          description: Token para refrescar el acceso
        tokenType:
          type: string
          description: Tipo de token (Bearer)
          default: Bearer
        expiresIn:
          type: integer
          format: int64
          description: Tiempo de expiración del token en segundos

    RefreshTokenRequestDTO:
      type: object
      required:
        - refreshToken
      properties:
        refreshToken:
          type: string
          description: Token de refresco

    ChangePasswordDTO:
      type: object
      required:
        - currentPassword
        - newPassword
      properties:
        currentPassword:
          type: string
          format: password
          description: Contraseña actual del usuario
        newPassword:
          type: string
          format: password
          description: Nueva contraseña del usuario (mínimo 8 caracteres, debe incluir mayúscula, minúscula, número y carácter especial)
          minLength: 8
          maxLength: 128
          pattern: '^(?=.*[a-z])(?=.*[A-Z])(?=.*\d)(?=.*[@$!%*?&])[A-Za-z\d@$!%*?&]{8,}$'

  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Token JWT de autenticación